changes select into radio buttons to speed up selections
having to click the select and then option required one step too much
and it was error prone as it could lead to selecting wrong option.

By having selectable options laid out as radio buttons, given the
options are just a handful of them, it makes things much faster
and less error prone.

If the number of options grows too much, it's necessary to implement
an autocomplete with suggestion in place.