adjusts success message when saving link
due to livewire, every time we save a link, the success message
was causing a jump in the page when it switched again while
staying there.

To avoid this problem, we need to absolute position the message
so it does not cause any jumps in the content, which is quite
annoying.

This change does exactly that. We no longer have the jump.