refactors code to use Laravel's Pipeline
by using Laravel's Pipeline we can pass an input through a series of
steps. Each preceding step return is the next step input.

This way, we have a clear view of what are the steps necessary to
accomplish certain behavior, in what order they happen and what
are their names.

The last bit, "their names", is one of the benefits of using Pipelines
as we have a direct view of classes names acting on the input.

The original class is not more concise, direct and easier to understand.
It also gives an easier way to maintain the code as each class encapsules
the necessary code for each step in a short amount of code.